Add runtime checks for multi-threading issues
Loadables were never supposed to be thread-safe and the default check for all
methods to be accessed on the main thread added here should be suitable for
most cases. Thus if you get an assertion after updating to this version most
likely there is a bug in the user code.

Still the update is considered breaking, because there are cases that would
require changes in the existing safe code:
- an object might be created on a worker thread (without touching shared data)
  and then used only from the main thread;
- an object might be accessed from non-main threads but exclusive access is
  enforced by another entity (wrapper).

In both cases the corresponding classes would need to override `+concurrency`
method.