Add checking availability for webp animation
There was an issue here that the webp format is supported from iOS 14 and higher. Therefore, a runtime error could appear if you run the code on iOS 13 and lower.
Also along the way, the minimum FPS was increased to 25, since some animations had it and looked very slow using previous value.